LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/animation-computed-expected.txt


FAIL Property animation value '1s' assert_equals: expected "1s ease 0s 1 normal none running none" but got ""
FAIL Property animation value 'cubic-bezier(0, -2, 1, 3)' assert_equals: expected "0s cubic-bezier(0, -2, 1, 3) 0s 1 normal none running none" but got ""
FAIL Property animation value '1s -3s' assert_equals: expected "1s ease -3s 1 normal none running none" but got ""
FAIL Property animation value '4' assert_equals: expected "0s ease 0s 4 normal none running none" but got ""
FAIL Property animation value 'reverse' assert_equals: expected "0s ease 0s 1 reverse none running none" but got ""
FAIL Property animation value 'both' assert_equals: expected "0s ease 0s 1 normal both running none" but got ""
FAIL Property animation value 'paused' assert_equals: expected "0s ease 0s 1 normal none paused none" but got ""
FAIL Property animation value 'none' assert_equals: expected "0s ease 0s 1 normal none running none" but got ""
FAIL Property animation value 'anim' assert_equals: expected "0s ease 0s 1 normal none running anim" but got ""
FAIL Property animation value 'anim paused both reverse 4 1s -3s cubic-bezier(0, -2, 1, 3)' assert_equals: expected "1s cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim" but got ""
FAIL Property animation value 'anim paused both reverse, 4 1s -3s cubic-bezier(0, -2, 1, 3)' assert_equals: expected "0s ease 0s 1 reverse both paused anim, 1s cubic-bezier(0, -2, 1, 3) -3s 4 normal none running none" but got ""
